+++ packages/chain-adapters/src/cosmossdk/CosmosSdkBaseAdapter.ts
@@ -240,6 +240,7 @@ export abstract class CosmosSdkBaseAdapter<T extends CosmosChainTypes> implement
         result: chainAdapters.ValidAddressResultType.Valid
       }
     } catch (err) {
+      console.error(err)
       return { valid: false, result: chainAdapters.ValidAddressResultType.Invalid }
     }
   }