+++ packages/swapper/src/swappers/test/TestSwapper.ts
@@ -31,11 +31,11 @@ export class TestSwapper implements Swapper {
     ]
   }
 
-  async getQuote(): Promise<Quote<ChainTypes, SwapperType>> {
+  async getQuote(): Promise<Quote<ChainTypes>> {
     throw new Error('TestSwapper: getQuote unimplemented')
   }
 
-  async buildQuoteTx(): Promise<Quote<ChainTypes, SwapperType>> {
+  async buildQuoteTx(): Promise<Quote<ChainTypes>> {
     throw new Error('TestSwapper: getQuote unimplemented')
   }
 
@@ -59,10 +59,6 @@ export class TestSwapper implements Swapper {
 
   async approveInfinite(): Promise<string> {
     throw new Error('TestSwapper: approveInfinite unimplemented')
-  }
-
-  async getSendMaxAmount(): Promise<string> {
-    throw new Error('TestSwapper: getSendMaxAmount unimplemented')
   }
 
   filterBuyAssetsBySellAssetId(args: BuyAssetBySellIdInput): CAIP19[] {