+++ packages/swapper/src/swappers/thorchain/ThorchainSwapper.ts
@@ -17,11 +17,11 @@ export class ThorchainSwapper implements Swapper {
     return SwapperType.Thorchain
   }
 
-  async getQuote(): Promise<Quote<ChainTypes, SwapperType>> {
+  async getQuote(): Promise<Quote<ChainTypes>> {
     throw new Error('ThorchainSwapper: getQuote unimplemented')
   }
 
-  async buildQuoteTx(): Promise<Quote<ChainTypes, SwapperType>> {
+  async buildQuoteTx(): Promise<Quote<ChainTypes>> {
     throw new Error('ThorchainSwapper: getQuote unimplemented')
   }
 
@@ -45,10 +45,6 @@ export class ThorchainSwapper implements Swapper {
 
   async approveInfinite(): Promise<string> {
     throw new Error('ThorchainSwapper: approveInfinite unimplemented')
-  }
-
-  async getSendMaxAmount(): Promise<string> {
-    throw new Error('ThorchainSwapper: getSendMaxAmount unimplemented')
   }
 
   filterBuyAssetsBySellAssetId(): CAIP19[] {