+++ packages/asset-service/CHANGELOG.md
@@ -1,3 +1,32 @@
+# [@shapeshiftoss/asset-service-v2.4.3](https://github.com/shapeshift/lib/compare/@shapeshiftoss/asset-service-v2.4.2...@shapeshiftoss/asset-service-v2.4.3) (2022-05-02)
+
+
+### Bug Fixes
+
+* removes 3 unnecessary api calls from zrxswapper.  ([#586](https://github.com/shapeshift/lib/issues/586)) ([57d59c8](https://github.com/shapeshift/lib/commit/57d59c8488cceb05cd5e0778b49690d0d2c97b68))
+
+# [@shapeshiftoss/asset-service-v2.4.2](https://github.com/shapeshift/lib/compare/@shapeshiftoss/asset-service-v2.4.1...@shapeshiftoss/asset-service-v2.4.2) (2022-05-02)
+
+
+### Bug Fixes
+
+* unbreak lib build ([#588](https://github.com/shapeshift/lib/issues/588)) ([bff0f3d](https://github.com/shapeshift/lib/commit/bff0f3d351f09ae9693b6b173782e8d8671ca3e4))
+* vscode tsconfig errors ([#577](https://github.com/shapeshift/lib/issues/577)) ([50138d0](https://github.com/shapeshift/lib/commit/50138d07b55b730f3bee68fae80414dc6578ee2a))
+
+# [@shapeshiftoss/asset-service-v2.4.1](https://github.com/shapeshift/lib/compare/@shapeshiftoss/asset-service-v2.4.0...@shapeshiftoss/asset-service-v2.4.1) (2022-04-29)
+
+
+### Bug Fixes
+
+* update assets file ([#583](https://github.com/shapeshift/lib/issues/583)) ([64a9572](https://github.com/shapeshift/lib/commit/64a957283a65635d9f3fae16bb8a12a95002dc7f))
+
+# [@shapeshiftoss/asset-service-v2.4.0](https://github.com/shapeshift/lib/compare/@shapeshiftoss/asset-service-v2.3.1...@shapeshiftoss/asset-service-v2.4.0) (2022-04-28)
+
+
+### Features
+
+* add assetid, chainid fields to asset type ([#570](https://github.com/shapeshift/lib/issues/570)) ([1c3c24c](https://github.com/shapeshift/lib/commit/1c3c24c2df6e71f3a4ad4b7f1863168aafdc8aa5))
+
 # [@shapeshiftoss/asset-service-v2.3.1](https://github.com/shapeshift/lib/compare/@shapeshiftoss/asset-service-v2.3.0...@shapeshiftoss/asset-service-v2.3.1) (2022-04-16)